modify to response setting in admin info, event, home.

+++ app/backend/app/Services/Admins/InformationsService.php
@@ -24,6 +24,7 @@ use App\Library\Cache\CacheLibrary;
 use App\Library\Time\TimeLibrary;
 use App\Models\Masters\Informations;
 use Exception;
+use Symfony\Component\HttpFoundation\BinaryFileResponse;
 
 class InformationsService
 {
@@ -47,9 +48,9 @@ class InformationsService
      * get information data
      *
      * @param
-     * @return JsonResponse
+     * @return array
      */
-    public function getInformations(): JsonResponse
+    public function getInformations(): array
     {
         $cache = CacheLibrary::getByKey(self::CACHE_KEY_INFORMATION_COIN_COLLECTION_LIST);
 
@@ -62,10 +63,10 @@ class InformationsService
                 CacheLibrary::setCache(self::CACHE_KEY_INFORMATION_COIN_COLLECTION_LIST, $resourceCollection);
             }
         } else {
-            $resourceCollection = $cache;
+            $resourceCollection = (array)$cache;
         }
 
-        return response()->json($resourceCollection, 200);
+        return $resourceCollection;
     }
 
     /**
@@ -73,7 +74,7 @@ class InformationsService
      *
      * @return \Symfony\Component\HttpFoundation\BinaryFileResponse
      */
-    public function downloadCSV()
+    public function downloadCSV(): BinaryFileResponse
     {
         $data = $this->informationsRepository->getRecords();
 
@@ -85,7 +86,7 @@ class InformationsService
      *
      * @return \Symfony\Component\HttpFoundation\BinaryFileResponse
      */
-    public function downloadTemplate()
+    public function downloadTemplate(): BinaryFileResponse
     {
         return Excel::download(
             new InformationsBulkInsertTemplateExport(collect(Config::get('myappFile.service.admins.informations.template'))),
@@ -98,9 +99,9 @@ class InformationsService
      * imort informations by template data service
      *
      * @param UploadedFile $file
-     * @return JsonResponse
+     * @return void
      */
-    public function importTemplate(UploadedFile $file)
+    public function importTemplate(UploadedFile $file): void
     {
         // ファイル名チェック
         if (!preg_match('/^master_informations_template_\d{14}\.xlsx/u', $file->getClientOriginalName())) {
@@ -119,18 +120,24 @@ class InformationsService
             // $resource = app()->make(GameEnemiesCreateResource::class, ['resource' => $fileData[0]])->toArray($request);
             $resource = InformationsResource::toArrayForBulkInsert(current($fileData));
 
-            $insertCount = $this->informationsRepository->create($resource);
+            $result = $this->informationsRepository->create($resource);
+
+            // 作成出来ない場合
+            if (!$result) {
+                throw new MyApplicationHttpException(
+                    StatusCodeMessages::STATUS_401,
+                    parameter: [
+                        'resource' => $resource,
+                    ]
+                );
+            }
 
             DB::commit();
 
             // キャッシュの削除
             CacheLibrary::deleteCache(self::CACHE_KEY_INFORMATION_COIN_COLLECTION_LIST, true);
 
-            // レスポンスの制御
-            $message = $insertCount ? 'success' : 'Bad Request';
-            $status = $insertCount ? 201 : 401;
-
-            return response()->json(['message' => $message, 'status' => $status], $status);
+            // return response()->json(['message' => $message, 'status' => $status], $status);
         } catch (Exception $e) {
             Log::error(__CLASS__ . '::' . __FUNCTION__ . ' line:' . __LINE__ . ' ' . 'message: ' . json_encode($e->getMessage()));
             DB::rollback();
@@ -146,26 +153,37 @@ class InformationsService
      * @param string $detail detail
      * @param string $startAt start datetime
      * @param string $endAt end datetime
-     * @return \Illuminate\Http\JsonResponse
+     * @return void
      */
-    public function createInformation(string $name, int $type, string $detail, string $startAt, string $endAt): JsonResponse
-    {
+    public function createInformation(
+        string $name,
+        int $type,
+        string $detail,
+        string $startAt,
+        string $endAt
+    ): void {
         $resource = InformationsResource::toArrayForCreate($name, $type, $detail, $startAt, $endAt);
 
         DB::beginTransaction();
         try {
-            $insertCount = $this->informationsRepository->create($resource);
+            $result = $this->informationsRepository->create($resource);
+
+            // 作成出来ない場合
+            if (!$result) {
+                throw new MyApplicationHttpException(
+                    StatusCodeMessages::STATUS_401,
+                    parameter: [
+                        'resource' => $resource,
+                    ]
+                );
+            }
 
             DB::commit();
 
             // キャッシュの削除
             CacheLibrary::deleteCache(self::CACHE_KEY_INFORMATION_COIN_COLLECTION_LIST, true);
 
-            // 作成されている場合は304
-            $message = ($insertCount > 0) ? 'success' : 'Bad Request';
-            $status = ($insertCount > 0) ? 201 : 401;
-
-            return response()->json(['message' => $message, 'status' => $status], $status);
+            // return response()->json(['message' => $message, 'status' => $status], $status);
         } catch (Exception $e) {
             Log::error(__CLASS__ . '::' . __FUNCTION__ . ' line:' . __LINE__ . ' ' . 'message: ' . json_encode($e->getMessage()));
             DB::rollback();
@@ -182,10 +200,16 @@ class InformationsService
      * @param string $detail detail
      * @param string $startAt start datetime
      * @param string $endAt end datetime
-     * @return \Illuminate\Http\JsonResponse
+     * @return void
      */
-    public function updateInformation(int $id, string $name, int $type, string $detail, string $startAt, string $endAt): JsonResponse
-    {
+    public function updateInformation(
+        int $id,
+        string $name,
+        int $type,
+        string $detail,
+        string $startAt,
+        string $endAt
+        ): void {
         $resource = InformationsResource::toArrayForUpdate($name, $type, $detail, $startAt, $endAt);
 
         DB::beginTransaction();
@@ -194,16 +218,23 @@ class InformationsService
             $information = $this->getInformationById($id);
             $updatedRowCount = $this->informationsRepository->update($information[Informations::ID], $resource);
 
+            // 更新出来ない場合
+            if (!($updatedRowCount > 0)) {
+                throw new MyApplicationHttpException(
+                    StatusCodeMessages::STATUS_401,
+                    parameter: [
+                        'resource' => $resource,
+                        'information.id' => $id,
+                    ]
+                );
+            }
+
             DB::commit();
 
             // キャッシュの削除
             CacheLibrary::deleteCache(self::CACHE_KEY_INFORMATION_COIN_COLLECTION_LIST, true);
 
-            // 更新されていない場合は304
-            $message = ($updatedRowCount > 0) ? 'success' : 'not modified';
-            $status = ($updatedRowCount > 0) ? 200 : 304;
-
-            return response()->json(['message' => $message, 'status' => $status], $status);
+            // return response()->json(['message' => $message, 'status' => $status], $status);
         } catch (Exception $e) {
             Log::error(__CLASS__ . '::' . __FUNCTION__ . ' line:' . __LINE__ . ' ' . 'message: ' . json_encode($e->getMessage()));
             DB::rollback();
@@ -215,9 +246,9 @@ class InformationsService
      * delete information data service
      *
      * @param array $informationIds id of records
-     * @return \Illuminate\Http\JsonResponse
+     * @return void
      */
-    public function deleteInformation(array $informationIds): JsonResponse
+    public function deleteInformation(array $informationIds): void
     {
         DB::beginTransaction();
         try {
@@ -228,16 +259,23 @@ class InformationsService
 
             $deleteRowCount = $this->informationsRepository->delete($informationIds, $resource);
 
+            // 削除出来ない場合
+            if (!($deleteRowCount > 0)) {
+                throw new MyApplicationHttpException(
+                    StatusCodeMessages::STATUS_401,
+                    parameter: [
+                        'resource' => $resource,
+                        'informationIds' => $informationIds,
+                    ]
+                );
+            }
+
             DB::commit();
 
             // キャッシュの削除
             CacheLibrary::deleteCache(self::CACHE_KEY_INFORMATION_COIN_COLLECTION_LIST, true);
 
-            // 更新されていない場合は304
-            $message = ($deleteRowCount > 0) ? 'success' : 'not deleted';
-            $status = ($deleteRowCount > 0) ? 200 : 401;
-
-            return response()->json(['message' => $message, 'status' => $status], $status);
+            // return response()->json(['message' => $message, 'status' => $status], $status);
         } catch (Exception $e) {
             Log::error(__CLASS__ . '::' . __FUNCTION__ . ' line:' . __LINE__ . ' ' . 'message: ' . json_encode($e->getMessage()));
             DB::rollback();
